Templating fix for when an escaped character is used in Javascript
If an escaped character (e.g. \n) is used in embedded Javascript inside
a template then it gets escaped to \\n. If someone has put an escaped
character in their JS then they probably want it left that way.

This commit replaces \\s with a single \ when inside evaluation tags,
reverting the previous template escaping.